hw12 code review: оставила MainActivity.kt общей для всех buildTypes, вынесла различающуюся для free и paid функцию в utils-ы
